removeItem implemented
